First pass on routes; also upgrade MZ.
Needed to upgrade MZ to get around a routing issue (MZ PR 84).
That got into some type changes around `RawHtmlEl` which pulled
`.children`. Replace the children work with individual child calls.
