+++ backend/Cargo.toml
@@ -10,5 +10,5 @@ readme = "../README.md"
 edition = "2021"
 
 [dependencies]
-moon = { git = "https://github.com/MoonZoon/MoonZoon", rev = "15cb619faca5f78a47e08f4af4bfa595f0eb64b1" }
+moon = { git = "https://github.com/MoonZoon/MoonZoon", rev = "5769c15d6376ce591120c994764809c1a65ed7bd" }
 shared = { path = "../shared", features = ["backend"] }
