+++ backend/src/main.rs
@@ -14,8 +14,8 @@ fn read_user_from_file<P: AsRef<Path>>(path: P) -> Result<Utterance, Box<dyn Err
     Ok(tx)
 }
 
-fn get_transcription_results(id: usize) -> Option<BlockMessage> {
-    let path = format!("./public/assets/block_{:04}.json", id);
+fn get_transcription_results(event_name: String, id: usize) -> Option<BlockMessage> {
+    let path = format!("./public/assets/{}/block_{:04}.json", event_name, id);
     match read_user_from_file(path) {
         Ok(block) => {
             let speaker = block.speaker.clone().unwrap_or_else(|| "".to_string());
@@ -47,19 +47,24 @@ async fn up_msg_handler(req: UpMsgRequest<UpMsg>) {
 
     match up_msg {
         UpMsg::DeleteBlock(block) => {
-            sessions::broadcast_down_msg(&DownMsg::BlockDeleted(block), cor_id).await;
+            println!("Delete Block{:?}", block.id);
+            // sessions::broadcast_down_msg(&DownMsg::BlockDeleted(block), cor_id).await;
         }
         UpMsg::EditBlock(block) => {
+            println!("Edit Block{:?}", block.id);
             sessions::broadcast_down_msg(&DownMsg::BlockEdited(block), cor_id).await;
         }
         UpMsg::MergeBlockAbove(block) => {
             sessions::broadcast_down_msg(&DownMsg::BlockMergedWithAbove(block), cor_id).await;
         }
         UpMsg::ChooseEvent(event) => {
+            println!("Choose event {}", event.id);
             let stream = EventStreamMessage {
                 id: event.id,
-                data: "Event 0001".to_string(),
+                data: format!("Selected event {}", event.id),
             };
+            let event_name = format!("event_{:04}", event.id);
+
             sessions::broadcast_down_msg(&DownMsg::EventSelected(stream), cor_id).await;
 
             static NEXT_ID: AtomicUsize = AtomicUsize::new(1);
@@ -67,7 +72,11 @@ async fn up_msg_handler(req: UpMsgRequest<UpMsg>) {
                 loop {
                     // We may not have the next file on disk, no worries, sleep and come back later
                     let id = NEXT_ID.load(Ordering::SeqCst);
-                    if let Some(block) = get_transcription_results(NEXT_ID.load(Ordering::SeqCst)) {
+                    if let Some(block) = get_transcription_results(
+                        event_name.clone(),
+                        NEXT_ID.load(Ordering::SeqCst),
+                    ) {
+                        println!("Loading file {:?}", id);
                         sessions::broadcast_down_msg(&DownMsg::BlockCreated(block), cor_id).await;
                         NEXT_ID.store(id + 1, Ordering::SeqCst);
                     }
