+++ shared/Cargo.toml
@@ -10,7 +10,7 @@ readme = "../README.md"
 edition = "2021"
 
 [dependencies]
-moonlight = { git = "https://github.com/MoonZoon/MoonZoon", rev = "15cb619faca5f78a47e08f4af4bfa595f0eb64b1" }
+moonlight = { git = "https://github.com/MoonZoon/MoonZoon", rev = "5769c15d6376ce591120c994764809c1a65ed7bd" }
 
 [features]
 frontend = ["moonlight/frontend"]
